# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Golang engineer test assessment - the Spy Cat Agency",
        "title": "Golang engineer test assessment - the Spy Cat Agency",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "basePath": "/v1",
    "paths": {
...
        "/ql/{id}": {
            "get": {
                "description": "List of cats",
                "tags": [
                    "spycat"
                ],
                "summary": "List of cats",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cat ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/spycat": {
            "get": {
                "description": "Fetches spy cat list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "spycat"
                ],
                "summary": "Fetches spy cat list",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/store.Cat"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "post": {
                "description": "Create new spy cat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "spycat"
                ],
                "summary": "Create spy cat",
                "parameters": [
                    {
                        "description": "SpyCat payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CreateCatPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/store.Cat"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/spycat/{id}": {
            "get": {
                "description": "Get cat info by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "spycat"
                ],
                "summary": "Get cat info",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cat ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.Cat"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Delete spy cat by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "spycat"
                ],
                "summary": "Delete spy cat",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cat ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "patch": {
                "description": "Update cat salary by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "spycat"
                ],
                "summary": "Update cat salary",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cat ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update SpyCat payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.UpdateCatInfoPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.Cat"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
        "main.CreateCatPayload": {
            "type": "object",
            "required": [
                "breed",
                "name",
                "salary",
                "year_of_experience"
            ],
            "properties": {
                "breed": {
                    "type": "string",
                    "maxLength": 200
                },
                "name": {
                    "type": "string",
                    "maxLength": 200
                },
                "salary": {
                    "type": "integer",
                    "minimum": 1
                },
                "year_of_experience": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
...
        "main.UpdateCatInfoPayload": {
            "type": "object",
            "required": [
                "salary"
            ],
            "properties": {
                "salary": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
...
        "store.Cat": {
            "type": "object",
            "properties": {
                "breed": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "salary": {
                    "type": "integer"
                },
                "year_of_experience": {
                    "type": "integer"
                }
            }
        },
...
    }
}